Move meta types to Meta namespace
