+++ src/Service/Lexicon/V1/DefGenerator/Primary/ProcedureGenerator.php
@@ -10,8 +10,8 @@ use Blugen\Service\Lexicon\V1\Factory\ComponentGeneratorFactory;
 use Blugen\Service\Lexicon\V1\Resolver\NamespaceResolver;
 use Blugen\Service\Lexicon\V1\Resolver\NsidResolver;
 use Blugen\Service\Lexicon\V1\Schema\Container\ObjectSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\RefSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\UnionSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\RefSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\UnionSchema;
 use Blugen\Service\Lexicon\V1\TypeSpecificDefinition\Primary\ProcedureTypeDefinition;
 use Blugen\Service\Syntax\Factory\SchemaFactory;
 use Blugen\Service\Xrpc\CallableInterface;
