+++ src/Service/Lexicon/V1/Schema/Meta/RefSchema.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\Schema\Field;
+namespace Blugen\Service\Lexicon\V1\Schema\Meta;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 use Blugen\Service\Lexicon\V1\Traits\ArrayableTrait;
 use Blugen\Service\Lexicon\V1\Traits\RawSchemaAccessorTrait;
 
-class UnknownSchema implements SchemaInterface
+class RefSchema implements SchemaInterface
 {
     use ArrayableTrait;
     use RawSchemaAccessorTrait;
 
-    public function __construct(private readonly SchemaInterface $schema)
-    {
-
-    }
+    public function __construct(
+        private readonly SchemaInterface $schema,
+    )
+    {}
 
     public function type(): string
     {
@@ -29,5 +29,10 @@ class UnknownSchema implements SchemaInterface
     public function __get(string $name): mixed
     {
         return $this->schema->__get($name);
+    }
+
+    public function ref(): string
+    {
+        return $this->__get('ref');
     }
 }
