+++ tests/Unit/Service/Lexicon/V1/TypeSpecificSchema/Field/UnionSchemaTest.php
@@ -3,7 +3,7 @@
 namespace Blugen\Tests\Unit\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\V1\Schema;
-use Blugen\Service\Lexicon\V1\Schema\Field\UnionSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\UnionSchema;
 use PHPUnit\Framework\TestCase;
 
 class UnionSchemaTest extends TestCase
