+++ tests/Unit/Service/Lexicon/V1/TypeSpecificSchema/Field/UnknownSchemaTest.php
@@ -3,7 +3,7 @@
 namespace Blugen\Tests\Unit\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\V1\Schema;
-use Blugen\Service\Lexicon\V1\Schema\Field\UnknownSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\UnknownSchema;
 use PHPUnit\Framework\TestCase;
 
 class UnknownSchemaTest extends TestCase
