+++ tests/Unit/Service/Syntax/Factory/SchemaFactoryTest.php
@@ -11,9 +11,9 @@ use Blugen\Service\Lexicon\V1\Schema\Concrete\NullSchema;
 use Blugen\Service\Lexicon\V1\Schema\Concrete\StringSchema;
 use Blugen\Service\Lexicon\V1\Schema\Container\ArraySchema;
 use Blugen\Service\Lexicon\V1\Schema\Container\ObjectSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\TokenSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\UnionSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\UnknownSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\TokenSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\UnionSchema;
+use Blugen\Service\Lexicon\V1\Schema\Meta\UnknownSchema;
 use Blugen\Service\Syntax\Factory\SchemaFactory;
 use Blugen\Tests\TestCase;
 use PHPUnit\Framework\Attributes\DataProvider;
